ebin/app: Add missing ipa_proto_ccm module
Commit 0c40c626caa95fd27be5f6af9b0b9f63edb23bb2 introduced a new
ipa_proto_ccm module, which was not added to the app module
list. Commit 515aa3343442e784602e23890672e40eb320aa00 then used a
function from this module. In interactive mode the module would be
loaded automatically so its omission was not detected in testing. In
embedded mode (used by default in relx releases) modules are loaded
statically on startup and must be explicitly listed.

Change-Id: I5b7338341a9e179b41b8f2f092efe9aaf83451bb